Create findWordInStringFromListItems
Suppose we have a language where we don’t have white spaces between words. Now, the task is to segment the sentence into words. For example: themanranafterit is actually “the man ran after it”. 
Write a function tokenize(str, dict): which takes string(sentence) and the dictionary and returns the tokens